change new content write token prefix from "tq__" to "tqw__" (#1574)
Temporary fix because the JS client accepts write tokens only if they
start with "tqw_".

Temporarily change the new "tq__" prefix to "tqw__" (2 underscores!).
This can be parsed and distinguished from the old "tqw_" (with single
underscore) correctly, because "_" is not a valid base58 character. Then
once the clients are all updated in a couple of months, we change the
prefix back to "tq__", but also allow "tqw__" (2 underscores) as an
alias for the new token format.