fix lookup of client IP from HTTP request caused by regression in gin
gin 1.7.0 added a feature to accept/refuse proxy headers depending on
the remote IP of the TCP connection. A fix of a data race in gin 1.7.1
made that feature inaccessible unless gin is started through gin.Run(),
which we don't do in the fabric. A PR to fix this has not been approved
for over a month (https://github.com/gin-gonic/gin/pull/2692).

Hence I implemented similar logic in our own helper, and consolidated
with an existing helper (that we used in addition to the gin-specific
functionality!). In addition, the client IP is cached in the gin context
in order to improve performance (multiple middleware handlers use the
client IP for logging, usage, geo lookups, redirect, etc.)