fix opening part writer: do not list dir for finding the part (#1891)
Non-finalized part files are created with the encryption scheme as
filename extension (the proper way would have been to encode the
encryption algorithm in the part write token as it was done with part
hashes...) The OpenPartWriter function globs the (temp) dir with in
order to find the part. However, listing/globbing a directory may not
find a part immediately after it was created due to filesystem delays/
caches. That problem was actually observed in live streaming.

To fix, try opening candidate files with the different extensions
directly.