prevent retrieval of all node locators on creation of new draft content (#2001)
Before we added the node ID to content write tokens, the client had to determine the ID of the current node (on which the new draft content was created) by retrieving the locators of all space nodes and finding the one corresponding to the client configuration. If the node was close, this was barely noticeable. If the node was distant, however, this could take a long time (e.g. over a minute with RTT 300ms, 120 nodes). This fix prevents the locator lookup if the node ID is available in the write token.