fix utc.UTC monotonic clock, improve PeriodicCall unit test (#1996)
time.Time keeps track of a "wall clock" for "time telling" as well as a "monotonic clock" for "time measurements" -
see documentation of the time package. The monotonic clock is automatically stripped from a Time instance that
results from a time operation (e.g. Add, Truncate) as well as timezone changes, unmarshalling, etc.

Since UTC changes the timezone from "local" to "UTC", this also strips the monotonic clock. However, we want to be
able to use UTC also for reliable time measurements... Hence we retain the original Time instance and use that for
time measurements.