add garbage collection for cached parts & GC metrics (#2012)
Add garbage collection for parts that do not belong to the storage partition(s) managed by the local node. It's configured through a new 'part_cache' section in the GC config:

  {
    "gc": {
      "enabled": true,
      "interval": "1h",
      "part_cache": {
        "high_watermark": 0.9,           // usage as fraction of total capacity [0.0, 1.0[
        "low_watermark": 0.85,
        "high_watermark_bytes": "3.8TB", // specific byte value - alternative to fractions above
        "low_watermark_bytes": "3.6TB"
      },
      "deletion": {
        "strategy": "move",
        "move_target_path": "/backup/deleted_parts",
        "apply_to_live_parts": false
      }
  }

When the high watermark for the part store's disk usage is reached, the garbage collector will start deleting cached parts until the low watermark is reached.

Watermarks can be defined as fraction of total storage capacity or as absolute byte values (with '_bytes' suffix). Bytes values take precedence if both are defined, but it's generally much clearer to just use one.

Parts are deleted in order of least recent access time of the part file, so the "part cache" essentially becomes an LRU cache.

In order to test GC (in general) more easily, a new 'deletion' section was also added to the GC config. It defines a deletion strategy with the following values:

* 'delete': regular deletion - simply removes the part/proof files
* 'move': moves or copies the part/proof files to a different filesytem location configured through 'move_target_path'. This path is ideally located on another storage volume, especially for GC of cached parts, since that is based on storage usage. If not specified, the path is the regular parts path with an appended '_removed' suffix, e.g. 'var/QDATA/PARTS_deleted'
* 'dry-run': does not delete or move the parts, simply logs a message.

Live parts are always deleted for real per default in order to remain backwards compatible. However, if `apply_to_liveParts` is set to true, live parts will be handled with the same deletion strategy as regular parts.

* ensure consistency of node accessibility check intervals

ensure inaccessibility interval does not exceed accessibility interval. This fixes distribution and live restart integration tests that started failing because the default value of the inaccessibility interval would artificially increase the accessibility interval that was specifically configured for these tests.

* metrics improvements

* code review