replace go.opentelemetry.io/otel with minimal custom tracing implementation
The version of go.opentelemetry.io/otel used is very old and incompatible with newer versions. When combined with other
modules that use opentelemetry, it almost certainly creates version conflicts. Upgrading the code to the latest version
of go.opentelemetry.io/otel is very complex and might not even be possible due to many incompatible changes of the APIs.