do not serve "very old" entries with ServeStaleDuringRefresh option
This addresses the case where an entry is created & cached, but then not requested for a long time. On next access, it should not be served as stale - instead it should be evicted and re-created, blocking the caller during construction. There are two cases: the entry was created through a refresh or regularly.
* do not serve stale entry if it was produced with a previous refresh, but has expired in the meantime
* do not serve stale entry if it is older than the entry's construction duration plus max age and no refresh is going on. Previously, we would kick off a refresh and serve the stale entry. Now we follow the regular "expiration path": evict the entry and block until new entry is constructed.

This commit also fixes metrics.Removed, which was incremented twice in some cases.